refactor : Member 패키지 경로 수정에 따른 리팩토링
