+++ Cockpybara/src/main/java/Alchole_free/Cockpybara/controller/member/detail/DetailResponse.java
@@ -1,6 +1,6 @@
 package Alchole_free.Cockpybara.controller.member.detail;
 
-import Alchole_free.Cockpybara.domain.Member;
+import Alchole_free.Cockpybara.domain.member.Member;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
