+++ Cockpybara/src/main/java/Alchole_free/Cockpybara/controller/member/update/MemberInfoUpdateResponse.java
@@ -1,6 +1,6 @@
 package Alchole_free.Cockpybara.controller.member.update;
 
-import Alchole_free.Cockpybara.domain.Member;
+import Alchole_free.Cockpybara.domain.member.Member;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
